Move TabularDataset to common
