+++ common/src/autogluon/common/__init__.py
@@ -1,3 +1,4 @@
+from .dataset import TabularDataset
 from .features.feature_metadata import FeatureMetadata
 from .utils.log_utils import _add_stream_handler
 from .utils.log_utils import fix_logging_if_kaggle as __fix_logging_if_kaggle
