+++ tabular/src/autogluon/tabular/__init__.py
@@ -1,6 +1,6 @@
+from autogluon.common.dataset import TabularDataset
 from autogluon.common.features.feature_metadata import FeatureMetadata
 from autogluon.common.utils.log_utils import _add_stream_handler
-from autogluon.core.dataset import TabularDataset
 
 try:
     from .version import __version__
